+++ backend/posts/views.py
@@ -2,6 +2,8 @@ from django.shortcuts import render, redirect
 from .models import Posts
 from .forms import PostsForm
 
+from django.contrib.auth.decorators import login_required
+
 from fruits.models import Fruits
 from commands.models import Commands
 from colors.models import Colors
@@ -19,12 +21,15 @@ def home(request):
     }
     return render(request, "home.html", context)
 
+@login_required
 def create_post(request):
 
     if request.method == "POST":
         form = PostsForm(request.POST)
         if form.is_valid():
-            form.save()
+            post = form.save(commit=False)
+            post.author = request.user
+            post.save()
         return redirect("home")
 
     context = {
@@ -32,21 +37,26 @@ def create_post(request):
     }
     return render(request, "create_post.html", context)
 
+@login_required
 def update_post(request, pk: int):
     post = Posts.objects.get(pk=pk)
+    if request.user.id == post.author.id:
+        if request.method == "POST":
+            form = PostsForm(request.POST, instance=post)
+            if form.is_valid():
+                form.save()
+            return redirect("home")
 
-    if request.method == "POST":
-        form = PostsForm(request.POST, instance=post)
-        if form.is_valid():
-            form.save()
+        context = {
+            "form": PostsForm(instance=post),
+        }
+        return render(request, "update_post.html", context)
+    else:
         return redirect("home")
 
-    context = {
-        "form": PostsForm(instance=post),
-    }
-    return render(request, "update_post.html", context)
-
+@login_required
 def delete_post(request, pk: int):
     post = Posts.objects.get(pk=pk)
-    post.delete()
+    if request.user.id == post.author.id:
+        post.delete()
     return redirect("home")
